backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docker==7.1.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
mcp==1.9.2
openai==1.82.1
openai-agents==0.0.16
outcome==1.3.0.post0
packaging==25.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pysocks==1.7.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.0.20250515
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
